First attempt to remove Sources
